move tests to actions test
