+++ tools/c7n_kube/tests/test_policy.py
@@ -2,8 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 from common_kube import KubeTest
 
-from c7n.exceptions import PolicyValidationError
-
 
 class TestAdmissionControllerMode(KubeTest):
     def test_kube_admission_policy(self):
@@ -158,136 +156,3 @@ class TestAdmissionControllerMode(KubeTest):
         result, resources = policy.push(event)
         self.assertEqual(result, 'deny')
         self.assertEqual(len(resources), 1)
-
-    def test_validator_event_label(self):
-        factory = self.replay_flight_data()
-        policy = self.load_policy(
-            {
-                'name': 'label-pod',
-                'resource': 'k8s.pod',
-                'mode': {
-                    'type': 'k8s-validator',
-                    'on-match': 'allow',
-                    'operations': ['CREATE']
-                },
-                'actions': [
-                    {
-                        'type': 'event-label',
-                        'labels': {
-                            'foo': 'bar',
-                            'role': 'different role',
-                            'test': None
-                        }
-                    }
-                ]
-
-            },
-            session_factory=factory,
-        )
-        event = self.get_event('create_pod')
-        result, resources = policy.push(event)
-        self.assertEqual(result, 'allow')
-        self.assertEqual(len(resources), 1)
-        self.assertEqual(len(resources[0]['c7n:patches']), 3)
-        self.assertEqual(
-            resources[0]['c7n:patches'],
-            [
-                {'op': 'remove', 'path': '/metadata/labels/test'},
-                {"op": "add", "path": "/metadata/labels/foo", "value": "bar"},
-                {"op": "replace", "path": "/metadata/labels/role", "value": "different role"},
-            ]
-        )
-
-    def test_validator_event_auto_label_user(self):
-        factory = self.replay_flight_data()
-        policy = self.load_policy(
-            {
-                'name': 'label-pod',
-                'resource': 'k8s.pod',
-                'mode': {
-                    'type': 'k8s-validator',
-                    'on-match': 'allow',
-                    'operations': ['CREATE']
-                },
-                'actions': [
-                    {
-                        'type': 'auto-label-user',
-                    }
-                ]
-
-            },
-            session_factory=factory,
-        )
-        event = self.get_event('create_pod')
-        result, resources = policy.push(event)
-        self.assertEqual(result, 'allow')
-        self.assertEqual(len(resources), 1)
-        self.assertEqual(len(resources[0]['c7n:patches']), 1)
-        self.assertEqual(
-            resources[0]['c7n:patches'],
-            [{"op": "add", "path": "/metadata/labels/OwnerContact", "value": "kubernetes-admin"}]
-        )
-
-    def test_validator_action_validate(self):
-        factory = self.replay_flight_data()
-        with self.assertRaises(PolicyValidationError):
-            self.load_policy(
-                {
-                    'name': 'label-pod',
-                    'resource': 'k8s.pod',
-                    'mode': {
-                        'type': 'k8s-validator',
-                        'on-match': 'allow',
-                        'operations': ['CREATE']
-                    },
-                    'actions': [
-                        {
-                            'type': 'label',
-                            'labels': {
-                                'foo': 'bar'
-                            }
-                        }
-                    ]
-
-                },
-                session_factory=factory,
-            )
-
-    def test_validator_action_event_patch(self):
-        factory = self.replay_flight_data()
-        policy = self.load_policy(
-            {
-                'name': 'label-pod',
-                'resource': 'k8s.pod',
-                'mode': {
-                    'type': 'k8s-validator',
-                    'on-match': 'allow',
-                    'operations': ['CREATE']
-                },
-                'actions': [
-                    {
-                        'type': 'event-patch',
-                        'key': 'spec.containers[].image',
-                        'value': '"prefix-{resource:kind}-"+.+"-{event:kind}"'
-                    }
-                ]
-
-            },
-            session_factory=factory,
-        )
-        event = self.get_event('create_pod')
-        result, resources = policy.push(event)
-        self.assertEqual(
-            resources[0]['c7n:patches'],
-            [
-                {
-                    'op': 'replace',
-                    'path': '/spec/containers/0/image',
-                    'value': 'prefix-Pod-ubuntu-AdmissionReview'},
-                {
-                    'op': 'replace',
-                    'path': '/spec/containers/1/image',
-                    'value': 'prefix-Pod-nginx-AdmissionReview'
-                }
-            ]
-        )
